Adding changes to modal search to bring items from backend
